First steps in adding unit tests
